fix: globally calc weight * item amount

+++ lib/app/data/models/character.dart
@@ -78,9 +78,9 @@ class Character with WithIcon implements WithMeta<Character, CharacterMeta> {
   int get maxXp => stats.maxXp;
   double get currentHpPercent => clamp(stats.currentHp / maxHp, 0, 1);
   double get currentExpPercent => clamp(stats.currentXp / maxXp, 0, 1);
-  int get maxLoad => stats.load ?? defaultMaxLoad;
-  int get defaultMaxLoad => characterClass.load + abilityScores.loadBaseValue;
-  int get currentLoad => items.fold(0, (weight, item) => weight + item.weight);
+  double get maxLoad => stats.load ?? defaultMaxLoad;
+  double get defaultMaxLoad => characterClass.load + abilityScores.loadBaseValue.toDouble();
+  double get currentLoad => items.fold(0, (weight, item) => weight + item.weight);
   int get armor => stats.armor ?? defaultArmor;
   int get defaultArmor => items.fold(0, (armor, item) => armor + item.armor);
   int get damageModifier => items.fold(0, (mod, item) => mod + item.damage);
